Next and previous buttons now enabled on spotify playback
In the NowPlaying view, the next and previous buttons were disabled when
using the spotify app. The lockscreen app did have a next button, and
that one worked.
I added a check on number of tracks, besides the check on isRemote. When
the current played song is "Remote" (internet radio or streaming service
like spotify), and the number of tracks is not 1, the previous and next
buttons are now enabled.
For radio stations the number of tracks is always one. For streaming
services it contains the number of tracks in the playlist.
The number of tracks was allready presentin the tokenMap, under the name
"playlist_tracks".